ENH: Respect activation
